Separate out processing of speculative and delegated NS RRsets.
The intention is that they can later have different error
handling semantics, as a speculative lookup should not fail
so hard - it is as if the cache was never populated.

Prior to this patch, only speculative lookups are made, with no
special treatment of delegated NSes; it is assumed that all
delegated NSes will be caught by those speculative lookups,
which I think is true, but crucially this mechanism does not
handle the separate error paths required.